Move InternalFlagOnlyMessage implementations from `.message` to `.message.flags`
